LayoutTests/fast/viewport/viewport-2-expected.txt

ALERT: viewport size 980x480 scale 0.733333 with limits [0.733333, 5] and userScalable true
